fix: made summary work more precisely + fixed/cleanup unit tests
