+++ src/Hl7.Fhir.Core.Tests/Serialization/SerializationTests.cs
@@ -128,406 +128,9 @@ namespace Hl7.Fhir.Tests.Serialization
             Assert.IsTrue(SerializationUtil.ProbeIsXml("<?xml />"));
         }
 
-        [TestMethod, Ignore] // Old tests, I'm note sure we need them anymore
-        public void TestSummary()
-        {
-            var p = new Patient
-            {
-                BirthDate = "1972-11-30",     // present in both summary and full
-                Photo = new List<Attachment>() { new Attachment() { ContentType = "text/plain" } }
-            };
-
-            var full = FhirXmlSerializer.SerializeToString(p);
-            Assert.IsTrue(full.Contains("<birthDate"));
-            Assert.IsTrue(full.Contains("<photo"));
-            Assert.IsNull(p.Meta, "Meta element should not be introduced here.");
-
-            var summ = FhirXmlSerializer.SerializeToString(p, summary: Fhir.Rest.SummaryType.True);
-            Assert.IsTrue(summ.Contains("<birthDate"));
-            Assert.IsFalse(summ.Contains("<photo"));
-            Assert.IsNull(p.Meta, "Meta element should not be introduced here.");
-
-            var q = new Questionnaire();
-            q.Text = new Narrative() { Div = "<div xmlns=\"http://www.w3.org/1999/xhtml\">Test Questionnaire</div>" };
-            q.Status = PublicationStatus.Active;
-            q.Date = "2015-09-27";
-            q.Title = "TITLE";
-            q.Item = new List<Questionnaire.ItemComponent>();
-            q.Item.Add(new Questionnaire.ItemComponent()
-            {
-                LinkId = "linkid",
-                Text = "TEXT"
-            });
-            
-            Assert.IsNull(q.Meta, "Meta element has not been created.");
-            var qfull = FhirXmlSerializer.SerializeToString(q);
-            Assert.IsNull(q.Meta, "Meta element should not be introduced here.");
-            Console.WriteLine("summary: Fhir.Rest.SummaryType.False");
-            Console.WriteLine(qfull);
-            Assert.IsTrue(qfull.Contains("Test Questionnaire"));
-            Assert.IsTrue(qfull.Contains("<status value=\"active\""));
-            Assert.IsTrue(qfull.Contains("<date value=\"2015-09-27\""));
-            Assert.IsTrue(qfull.Contains("<title value=\"TITLE\""));
-            Assert.IsTrue(qfull.Contains("<text value=\"TEXT\""));
-            Assert.IsTrue(qfull.Contains("<linkId value=\"linkid\""));
-
-            var qSum = FhirXmlSerializer.SerializeToString(q, summary: Fhir.Rest.SummaryType.True);
-            Console.WriteLine("summary: Fhir.Rest.SummaryType.True");
-            Console.WriteLine(qSum);
-            Assert.IsFalse(qSum.Contains("Test Questionnaire"));
-            Assert.IsTrue(qSum.Contains("<status value=\"active\""));
-            Assert.IsTrue(qSum.Contains("<date value=\"2015-09-27\""));
-            Assert.IsTrue(qSum.Contains("<title value=\"TITLE\""));
-            Assert.IsFalse(qSum.Contains("<text value=\"TEXT\""));
-            Assert.IsFalse(qSum.Contains("<linkId value=\"linkid\""));
-
-            var qData = FhirXmlSerializer.SerializeToString(q, summary: Fhir.Rest.SummaryType.Data);
-            Console.WriteLine("summary: Fhir.Rest.SummaryType.Data");
-            Console.WriteLine(qData);
-            Assert.IsFalse(qData.Contains("Test Questionnaire"));
-            Assert.IsTrue(qData.Contains("<meta"));
-            Assert.IsTrue(qData.Contains("<text value=\"TEXT\""));
-            Assert.IsTrue(qData.Contains("<status value=\"active\""));
-            Assert.IsTrue(qData.Contains("<date value=\"2015-09-27\""));
-            Assert.IsTrue(qData.Contains("<title value=\"TITLE\""));
-            Assert.IsTrue(qData.Contains("<linkId value=\"linkid\""));
-
-            q.Meta = new Meta { VersionId = "v2" };
-            var qText = FhirXmlSerializer.SerializeToString(q, summary: Fhir.Rest.SummaryType.Text);
-            Console.WriteLine("summary: Fhir.Rest.SummaryType.Text");
-            Console.WriteLine(qText);
-            Assert.IsTrue(qText.Contains("Test Questionnaire"));
-            Assert.IsTrue(qText.Contains("<meta"));
-            Assert.IsTrue(qText.Contains("<status value=\"active\""));
-            Assert.IsFalse(qText.Contains("<text value=\"TEXT\""));
-            Assert.IsFalse(qText.Contains("<date value=\"2015-09-27\""));
-            Assert.IsFalse(qText.Contains("<title value=\"TITLE\""));
-            Assert.IsFalse(qText.Contains("<linkId value=\"linkid\""));
-            Assert.AreEqual(0, q.Meta.Tag.Where(t => t.System == "http://hl7.org/fhir/v3/ObservationValue" && t.Code == "SUBSETTED").Count(), "Subsetted Tag should not still be there.");
-
-            // Verify that reloading the content into an object...
-            // make sure we accept the crappy output with empty groups
-            var nav = FhirXmlNode.Parse(qText, new FhirXmlParsingSettings { PermissiveParsing = true });
-
-            var qInflate = FhirXmlParser.Parse<Questionnaire>(nav);
-            Assert.AreEqual(1, qInflate.Meta.Tag.Where(t => t.System == "http://hl7.org/fhir/v3/ObservationValue" && t.Code == "SUBSETTED").Count(), "Subsetted Tag should not still be there.");
-        }
-
-        [TestMethod]
-        public void TestIncludeMandatory()
-        {
-            var l = new Library();
-            l.Type = new CodeableConcept { TextElement = new FhirString("testMandatoryComplexType") };
-            l.Id = "testId";
-            l.Language = "testLang";
-            var summaryElements = FhirXmlSerializer.SerializeToString(l, Fhir.Rest.SummaryType.Count);
-            
-            Assert.IsFalse(summaryElements.Contains("<language"));
-            Assert.IsTrue(summaryElements.Contains("<type>"));
-            Assert.IsTrue(summaryElements.Contains("<id value=\"testId\""));
-            
-            var customMaskingNode = new MaskingNode(new ScopedNode(l.ToTypedElement()), new MaskingNodeSettings
-            {
-                IncludeMandatory = true,
-                PreserveBundle = MaskingNodeSettings.PreserveBundleMode.All
-            });
-
-            var result = customMaskingNode.ToXml(settings: new FhirXmlSerializationSettings());
-
-            Assert.IsFalse(result.Contains("<language>"));
-            Assert.IsTrue(result.Contains("<type>"));
-            Assert.IsFalse(result.Contains("<id value=\"testId\""));
-
-            var b = new Bundle
-            {
-                TypeElement = new Code<Bundle.BundleType> { Value = Bundle.BundleType.Collection },
-                Entry = new List<Bundle.EntryComponent>()
-                {
-                    new Bundle.EntryComponent { Resource = l }
-                },
-                Id = "bundle-id"
-            };
-
-            var customMaskingNodeForBundle = new MaskingNode(new ScopedNode(b.ToTypedElement()), new MaskingNodeSettings
-            {
-                IncludeMandatory = true,
-                PreserveBundle = MaskingNodeSettings.PreserveBundleMode.None
-            });
-
-            result = customMaskingNodeForBundle.ToXml(settings: new FhirXmlSerializationSettings());
-            
-            Assert.IsTrue(result.Contains("<type value=\"collection\""));
-            Assert.IsFalse(result.Contains("<id value=\"bundle-id\""));
-        }
-
-        [TestMethod]
-        public void TestElements()
-        {
-            var p = new Patient
-            {
-                BirthDate = "1972-11-30",
-                Photo = new List<Attachment>() { new Attachment() { ContentType = "text/plain" } }
-            };
-            var elements = new[] { "photo" };
-            
-            var summaryElements = FhirXmlSerializer.SerializeToString(p, Fhir.Rest.SummaryType.False, elements: elements);
-            Assert.IsFalse(summaryElements.Contains("<birthDate"));
-            Assert.IsTrue(summaryElements.Contains("<photo"));
-
-            var noSummarySpecified = FhirXmlSerializer.SerializeToString(p, elements: elements);
-            Assert.IsFalse(noSummarySpecified.Contains("<birthDate"));
-            Assert.IsTrue(noSummarySpecified.Contains("<photo"));
-
-            Assert.ThrowsException<ArgumentException>(() => FhirXmlSerializer.SerializeToString(p, Fhir.Rest.SummaryType.True, elements: elements));
-            Assert.ThrowsException<ArgumentException>(() => FhirXmlSerializer.SerializeToString(p, Fhir.Rest.SummaryType.Count, elements: elements));
-            Assert.ThrowsException<ArgumentException>(() => FhirXmlSerializer.SerializeToString(p, Fhir.Rest.SummaryType.Data, elements: elements));
-            Assert.ThrowsException<ArgumentException>(() => FhirXmlSerializer.SerializeToString(p, Fhir.Rest.SummaryType.Text, elements: elements));
-        }
-
-        [TestMethod]
-        public void TestWithMetadata()
-        {
-            var p = new Patient
-            {
-                BirthDate = "1972-11-30"
-            };
-
-            var pSum = FhirXmlSerializer.SerializeToString(p, summary: Fhir.Rest.SummaryType.True);
-            Assert.IsNull(p.Meta, "Meta should not be there");
-
-            p.Meta = new Meta { VersionId = "v2" }; // introducing meta data ourselves. 
-
-            pSum = FhirXmlSerializer.SerializeToString(p, summary: Fhir.Rest.SummaryType.True);
-            Assert.IsNotNull(p.Meta, "Meta should still be there");
-            Assert.AreEqual(0, p.Meta.Tag.Where(t => t.System == "http://hl7.org/fhir/v3/ObservationValue" && t.Code == "SUBSETTED").Count(), "Subsetted Tag should not still be there.");
-        }
-
         private FhirXmlParser FhirXmlParser = new FhirXmlParser();
         private FhirJsonParser FhirJsonParser = new FhirJsonParser();
 
-        [TestMethod]
-        public void TestBundleSummary()
-        {
-            var p = new Patient
-            {
-                BirthDate = "1972-11-30",     // present in both summary and full
-                Photo = new List<Attachment>() { new Attachment() { ContentType = "text/plain" } }
-            };
-
-            var b = new Bundle();
-            b.AddResourceEntry(p, "http://nu.nl/fhir/Patient/1");
-            b.Total = 1;
-            b.Type = Bundle.BundleType.Searchset;
-
-            var full = FhirXmlSerializer.SerializeToString(b);
-            Assert.IsTrue(full.Contains("<entry"));
-            Assert.IsTrue(full.Contains("<birthDate"));
-            Assert.IsTrue(full.Contains("<photo"));
-            Assert.IsTrue(full.Contains("<total"));
-
-            var summ = FhirXmlSerializer.SerializeToString(b, summary: Fhir.Rest.SummaryType.True);
-            Assert.IsTrue(summ.Contains("<entry"));
-            Assert.IsTrue(summ.Contains("<birthDate"));
-            Assert.IsFalse(summ.Contains("<photo"));
-            Assert.IsTrue(summ.Contains("<total"));
-
-            summ = FhirXmlSerializer.SerializeToString(b, summary: Fhir.Rest.SummaryType.Count);
-            Assert.IsFalse(summ.Contains("<entry"));
-            Assert.IsFalse(summ.Contains("<birthDate"));
-            Assert.IsFalse(summ.Contains("<photo"));
-            Assert.IsTrue(summ.Contains("<total"));
-            Assert.IsTrue(summ.Contains("<type"));
-        }
-
-
-        [TestMethod]
-        public void TestBundleWithSummaryJson()
-        {
-            var patientOne = new Patient
-            {
-
-                Id = "patient-one",
-                Text = new Narrative { Div = "<div xmlns='http://www.w3.org/1999/xhtml'>A great blues player</div>" },
-                Meta = new Meta { VersionId = "eric-clapton" },
-
-                Name = new List<HumanName> { new HumanName { Family = "Clapton", Use = HumanName.NameUse.Official } },
-
-                Active = true,
-                BirthDate = "2015-07-09",
-                Gender = AdministrativeGender.Male
-            };
-
-            var patientTwo = new Patient()
-            {
-                Id = "patient-two",
-                Active = true,
-                Text = new Narrative { Div = "<div xmlns='http://www.w3.org/1999/xhtml'>Another great blues player</div>" },
-                Meta = new Meta { VersionId = "bb-king" },
-                Name = new List<HumanName> { new HumanName { Family = "King", Use = HumanName.NameUse.Nickname } }
-            };
-
-            var bundle = new Bundle()
-            {
-                Id = "my-bundle",
-                Total = 1803,
-                Type = Bundle.BundleType.Searchset,
-                Entry = new List<Bundle.EntryComponent> {
-                    new Bundle.EntryComponent { Resource = patientOne, FullUrl = "http://base/Patient/patient-one", Search = new Bundle.SearchComponent() { Mode = Bundle.SearchEntryMode.Match } },
-                    new Bundle.EntryComponent { Resource = patientTwo, FullUrl = "http://base/Patient/patient-two", Search = new Bundle.SearchComponent() { Mode = Bundle.SearchEntryMode.Match } }
-                }
-            };
-
-            var trueBundle = FhirJsonSerializer.SerializeToString(bundle, Fhir.Rest.SummaryType.True);
-            var dataBundle = FhirJsonSerializer.SerializeToString(bundle, Fhir.Rest.SummaryType.Data);
-            var textBundle = FhirJsonSerializer.SerializeToString(bundle, Fhir.Rest.SummaryType.Text);
-            var countBundle = FhirJsonSerializer.SerializeToString(bundle, Fhir.Rest.SummaryType.Count);
-            var falseBundle = FhirJsonSerializer.SerializeToString(bundle, Fhir.Rest.SummaryType.False);
-
-            var shouldBeSummaryTrue = TestDataHelper.ReadTestData("summary\\bundle-summary-true.json");
-            var shouldBeSummaryData = TestDataHelper.ReadTestData("summary\\bundle-summary-data.json");
-            var shouldBeSummaryText = TestDataHelper.ReadTestData("summary\\bundle-summary-text.json");
-            var shouldBeSummaryCount = TestDataHelper.ReadTestData("summary\\bundle-summary-count.json");
-            var shouldBeSummaryFalse = TestDataHelper.ReadTestData("summary\\bundle-summary-false.json");
-
-            Assert.AreEqual(shouldBeSummaryTrue, trueBundle);
-            Assert.AreEqual(shouldBeSummaryData, dataBundle);
-            Assert.AreEqual(shouldBeSummaryText, textBundle);
-            Assert.AreEqual(shouldBeSummaryCount, countBundle);
-            Assert.AreEqual(shouldBeSummaryFalse, falseBundle);
-        }
-
-        [TestMethod]
-        public void TestBundleWithSummaryXML()
-        {
-            var patientOne = new Patient
-            {
-                Id = "patient-one",
-                Text = new Narrative { Div = "<div xmlns='http://www.w3.org/1999/xhtml'>A great blues player</div>", Status = Narrative.NarrativeStatus.Additional },
-                Meta = new Meta { VersionId = "eric-clapton" },
-
-                Name = new List<HumanName> { new HumanName { Family = "Clapton", Use = HumanName.NameUse.Official } },
-
-                Active = true,
-                BirthDate = "2015-07-09",
-                Gender = AdministrativeGender.Male
-            };
-
-            var patientTwo = new Patient()
-            {
-                Id = "patient-two",
-                Active = true,
-                Text = new Narrative { Div = "<div xmlns='http://www.w3.org/1999/xhtml'>Another great blues player</div>", Status = Narrative.NarrativeStatus.Additional },
-                Meta = new Meta { VersionId = "bb-king" },
-                Name = new List<HumanName> { new HumanName { Family = "King", Use = HumanName.NameUse.Nickname } }
-            };
-
-            var bundle = new Bundle()
-            {
-                Id = "my-bundle",
-                Total = 1803,
-                Type = Bundle.BundleType.Searchset,
-                Entry = new List<Bundle.EntryComponent> {
-                    new Bundle.EntryComponent { Resource = patientOne, FullUrl = "http://base/Patient/patient-one", Search = new Bundle.SearchComponent() { Mode = Bundle.SearchEntryMode.Match } },
-                    new Bundle.EntryComponent { Resource = patientTwo, FullUrl = "http://base/Patient/patient-two", Search = new Bundle.SearchComponent() { Mode = Bundle.SearchEntryMode.Match } }
-                }
-            };
-
-            var textBundle = FhirXmlSerializer.SerializeToString(bundle, Fhir.Rest.SummaryType.Text);
-            var dataBundle = FhirXmlSerializer.SerializeToString(bundle, Fhir.Rest.SummaryType.Data);
-            var countBundle = FhirXmlSerializer.SerializeToString(bundle, Fhir.Rest.SummaryType.Count);
-            var trueBundle = FhirXmlSerializer.SerializeToString(bundle, Fhir.Rest.SummaryType.True);
-            var falseBundle = FhirXmlSerializer.SerializeToString(bundle, Fhir.Rest.SummaryType.False);
-
-            var shouldBeSummaryText = TestDataHelper.ReadTestData("summary\\bundle-summary-text.xml");
-            var shouldBeSummaryData = TestDataHelper.ReadTestData("summary\\bundle-summary-data.xml");
-            var shouldBeSummaryCount = TestDataHelper.ReadTestData("summary\\bundle-summary-count.xml");
-            var shouldBeSummaryTrue = TestDataHelper.ReadTestData("summary\\bundle-summary-true.xml");
-            var shouldBeSummaryFalse = TestDataHelper.ReadTestData("summary\\bundle-summary-false.xml");
-
-            Assert.AreEqual(shouldBeSummaryFalse, falseBundle);
-            Assert.AreEqual(shouldBeSummaryTrue, trueBundle);
-            Assert.AreEqual(shouldBeSummaryData, dataBundle);
-            Assert.AreEqual(shouldBeSummaryCount, countBundle);
-            Assert.AreEqual(shouldBeSummaryText, textBundle);
-
-        }
-
-        [TestMethod]
-        public void TestResourceWithSummaryJson()
-        {
-            var patientOne = new Patient
-            {
-
-                Id = "patient-one",
-                Text = new Narrative { Div = "<div xmlns='http://www.w3.org/1999/xhtml'>A great blues player</div>" },
-                Meta = new Meta { VersionId = "1234" },
-
-                Name = new List<HumanName> { new HumanName { Family = "Clapton", Use = HumanName.NameUse.Official } },
-
-                Active = true,
-                BirthDate = "2015-07-09",
-                Gender = AdministrativeGender.Male
-            };
-
-            // Properties with IsSummary == true -> Id, Meta, Active, BirthDate, Gender, Name
-
-            var summaryTrue = FhirJsonSerializer.SerializeToString(patientOne, Fhir.Rest.SummaryType.True);
-            var summaryText = FhirJsonSerializer.SerializeToString(patientOne, Fhir.Rest.SummaryType.Text);
-            var summaryData = FhirJsonSerializer.SerializeToString(patientOne, Fhir.Rest.SummaryType.Data);
-            var summaryFalse = FhirJsonSerializer.SerializeToString(patientOne, Fhir.Rest.SummaryType.False);
-            /* It doesn't make sense to use SummaryType.Count on a single resource hence why I'm not testing it here. */
-
-            var shouldBePatientOneTrue =
-                "{\"resourceType\":\"Patient\",\"id\":\"patient-one\",\"meta\":{\"versionId\":\"1234\",\"tag\":[{\"system\":\"http://hl7.org/fhir/v3/ObservationValue\",\"code\":\"SUBSETTED\"}]},\"active\":true,\"name\":[{\"use\":\"official\",\"family\":\"Clapton\"}],\"gender\":\"male\",\"birthDate\":\"2015-07-09\"}";
-
-            var shouldBePatientOneText =
-                "{\"resourceType\":\"Patient\",\"id\":\"patient-one\",\"meta\":{\"versionId\":\"1234\",\"tag\":[{\"system\":\"http://hl7.org/fhir/v3/ObservationValue\",\"code\":\"SUBSETTED\"}]},\"text\":{\"div\":\"<div xmlns='http://www.w3.org/1999/xhtml'>A great blues player</div>\"}}";
-
-            var shouldBePationeOneData =
-                "{\"resourceType\":\"Patient\",\"id\":\"patient-one\",\"meta\":{\"versionId\":\"1234\",\"tag\":[{\"system\":\"http://hl7.org/fhir/v3/ObservationValue\",\"code\":\"SUBSETTED\"}]},\"active\":true,\"name\":[{\"use\":\"official\",\"family\":\"Clapton\"}],\"gender\":\"male\",\"birthDate\":\"2015-07-09\"}";
-
-            var shouldBePatientOneFalse = "{\"resourceType\":\"Patient\",\"id\":\"patient-one\",\"meta\":{\"versionId\":\"1234\"},\"text\":{\"div\":\"<div xmlns='http://www.w3.org/1999/xhtml'>A great blues player</div>\"},\"active\":true,\"name\":[{\"use\":\"official\",\"family\":\"Clapton\"}],\"gender\":\"male\",\"birthDate\":\"2015-07-09\"}";
-
-            Assert.AreEqual(summaryTrue, shouldBePatientOneTrue);
-            Assert.AreEqual(summaryText, shouldBePatientOneText);
-            Assert.AreEqual(summaryData, shouldBePationeOneData);
-            Assert.AreEqual(summaryFalse, shouldBePatientOneFalse);
-        }
-
-        [TestMethod]
-        public void TestResourceWithSummaryXML()
-        {
-            var patientOne = new Patient
-            {
-
-                Id = "patient-one",
-                Text = new Narrative { Status = Narrative.NarrativeStatus.Generated, Div = "<div xmlns='http://www.w3.org/1999/xhtml'>A great blues player</div>" },
-                Meta = new Meta { ElementId = "eric-clapton", VersionId = "1234" },
-
-                Name = new List<HumanName> { new HumanName { Family = "Clapton", Use = HumanName.NameUse.Official } },
-
-                Active = true,
-                BirthDate = "2015-07-09",
-                Gender = AdministrativeGender.Male
-            };
-
-            // Properties with IsSummary == true -> Id, Meta, Active, BirthDate, Gender, Name
-
-            var summaryTrue = FhirXmlSerializer.SerializeToString(patientOne, Fhir.Rest.SummaryType.True);
-            var summaryText = FhirXmlSerializer.SerializeToString(patientOne, Fhir.Rest.SummaryType.Text);
-            var summaryData = FhirXmlSerializer.SerializeToString(patientOne, Fhir.Rest.SummaryType.Data);
-            var summaryFalse = FhirXmlSerializer.SerializeToString(patientOne, Fhir.Rest.SummaryType.False);
-
-            var shouldBeSummaryTrue = TestDataHelper.ReadTestData("summary\\summary-true.xml");
-            var shouldBeSummaryText = TestDataHelper.ReadTestData("summary\\summary-text.xml");
-            var shouldBeSummaryData = TestDataHelper.ReadTestData("summary\\summary-data.xml");
-            var shouldBeSummaryFalse = TestDataHelper.ReadTestData("summary\\summary-false.xml");
-
-            Assert.AreEqual(shouldBeSummaryTrue, summaryTrue);
-            Assert.AreEqual(shouldBeSummaryText, summaryText);
-            Assert.AreEqual(shouldBeSummaryData, summaryData);
-            Assert.AreEqual(shouldBeSummaryFalse, summaryFalse);
-        }
 
         //[TestMethod]
         //public void HandleCommentsJson()
@@ -582,67 +185,6 @@ namespace Hl7.Fhir.Tests.Serialization
             b = FhirXmlParser.Parse<Bundle>(xml);
 
             Assert.IsTrue(!b.NextLink.ToString().EndsWith("/"));
-        }
-
-        [TestMethod]
-        public void TestIdInSummary()
-        {
-            var p = new Patient
-            {
-                Id = "test-id-1",
-                BirthDate = "1972-11-30",     // present in both summary and full
-                Photo = new List<Attachment>() { new Attachment() { ContentType = "text/plain", Creation = "45" } },
-                ManagingOrganization = new ResourceReference() { Display = "temp org", Reference = "#temp" },
-
-                Text = new Narrative
-                {
-                    Div = "<div xmlns=\"http://www.w3.org/1999/xhtml\">Some test narrative</div>"
-                },
-                Meta = new Meta(),
-                Contained = new List<Resource>
-                {
-                    new Organization() { Id = "temp", Name = "temp org", Active = true }
-                }
-            };
-
-            p.AddExtension("http://example.org/ext", new FhirString("dud"));
-
-            var full = FhirXmlSerializer.SerializeToString(p);
-            Assert.IsTrue(full.Contains("narrative"));
-            Assert.IsTrue(full.Contains("dud"));
-            Assert.IsTrue(full.Contains("temp org"));
-            Assert.IsTrue(full.Contains("<id value="));
-            Assert.IsTrue(full.Contains("<birthDate"));
-            Assert.IsTrue(full.Contains("<photo"));
-            Assert.IsTrue(full.Contains("text/plain"));
-
-            full = FhirXmlSerializer.SerializeToString(p, summary: Hl7.Fhir.Rest.SummaryType.False);
-            Assert.IsTrue(full.Contains("narrative"));
-            Assert.IsTrue(full.Contains("dud"));
-            Assert.IsTrue(full.Contains("temp org"));
-            Assert.IsTrue(full.Contains("contain"));
-            Assert.IsTrue(full.Contains("<id value="));
-            Assert.IsTrue(full.Contains("<birthDate"));
-            Assert.IsTrue(full.Contains("<photo"));
-            Assert.IsTrue(full.Contains("text/plain"));
-
-            var summ = FhirXmlSerializer.SerializeToString(p, summary: Fhir.Rest.SummaryType.True);
-            Assert.IsFalse(summ.Contains("narrative"));
-            Assert.IsFalse(summ.Contains("dud"));
-            Assert.IsFalse(summ.Contains("contain"));
-            Assert.IsTrue(summ.Contains("temp org"));
-            Assert.IsTrue(summ.Contains("<id value="));
-            Assert.IsTrue(summ.Contains("<birthDate"));
-            Assert.IsFalse(summ.Contains("<photo"));
-
-            var data = FhirXmlSerializer.SerializeToString(p, summary: Hl7.Fhir.Rest.SummaryType.Data);
-            Assert.IsFalse(data.Contains("narrative"));
-            Assert.IsTrue(data.Contains("contain"));
-            Assert.IsTrue(data.Contains("dud"));
-            Assert.IsTrue(data.Contains("temp org"));
-            Assert.IsTrue(data.Contains("<id value="));
-            Assert.IsTrue(data.Contains("<birthDate"));
-            Assert.IsTrue(data.Contains("<photo"));
         }
 
         [TestMethod]
