+++ src/Hl7.Fhir.Core.Tests/TestData/summary/bundle-summary-false.json
@@ -1,1 +1,1 @@
-{"resourceType":"Bundle","id":"my-bundle","type":"searchset","total":1803,"entry":[{"fullUrl":"http://base/Patient/patient-one","resource":{"resourceType":"Patient","id":"patient-one","meta":{"versionId":"eric-clapton"},"text":{"div":"<div xmlns='http://www.w3.org/1999/xhtml'>A great blues player</div>"},"active":true,"name":[{"use":"official","family":"Clapton"}],"gender":"male","birthDate":"2015-07-09"},"search":{"mode":"match"}},{"fullUrl":"http://base/Patient/patient-two","resource":{"resourceType":"Patient","id":"patient-two","meta":{"versionId":"bb-king"},"text":{"div":"<div xmlns='http://www.w3.org/1999/xhtml'>Another great blues player</div>"},"active":true,"name":[{"use":"nickname","family":"King"}]},"search":{"mode":"match"}}]}
+{"resourceType":"Bundle","id":"my-bundle","type":"searchset","total":1803,"entry":[{"fullUrl":"http://base/Patient/patient-one","resource":{"resourceType":"Patient","id":"patient-one","meta":{"versionId":"eric-clapton"},"text":{"div":"<div xmlns='http://www.w3.org/1999/xhtml'>A great blues player</div>"},"active":true,"name":[{"use":"official","family":"Clapton"}],"gender":"male","birthDate":"2015-07-09"},"search":{"mode":"match"}},{"fullUrl":"http://base/Patient/patient-two","resource":{"resourceType":"Patient","id":"patient-two","meta":{"versionId":"bb-king"},"text":{"status":"additional","div":"<div xmlns='http://www.w3.org/1999/xhtml'>Another great blues player</div>"},"active":true,"name":[{"use":"nickname","family":"King"}]},"search":{"mode":"match"}}]}
