+++ src/Hl7.Fhir.Core.Tests/TestData/summary/bundle-summary-false.xml
@@ -1,1 +1,1 @@
-<Bundle xmlns="http://hl7.org/fhir"><id value="my-bundle" /><type value="searchset" /><total value="1803" /><entry><fullUrl value="http://base/Patient/patient-one" /><resource><Patient><id value="patient-one" /><meta><versionId value="eric-clapton" /></meta><text><status value="additional" /><div xmlns="http://www.w3.org/1999/xhtml">A great blues player</div></text><active value="true" /><name><use value="official" /><family value="Clapton" /></name><gender value="male" /><birthDate value="2015-07-09" /></Patient></resource><search><mode value="match" /></search></entry><entry><fullUrl value="http://base/Patient/patient-two" /><resource><Patient><id value="patient-two" /><meta><versionId value="bb-king" /></meta><text><status value="additional" /><div xmlns="http://www.w3.org/1999/xhtml">Another great blues player</div></text><active value="true" /><name><use value="nickname" /><family value="King" /></name></Patient></resource><search><mode value="match" /></search></entry></Bundle>
+<Bundle xmlns="http://hl7.org/fhir"><id value="my-bundle" /><type value="searchset" /><total value="1803" /><entry><fullUrl value="http://base/Patient/patient-one" /><resource><Patient><id value="patient-one" /><meta><versionId value="eric-clapton" /></meta><text><div xmlns="http://www.w3.org/1999/xhtml">A great blues player</div></text><active value="true" /><name><use value="official" /><family value="Clapton" /></name><gender value="male" /><birthDate value="2015-07-09" /></Patient></resource><search><mode value="match" /></search></entry><entry><fullUrl value="http://base/Patient/patient-two" /><resource><Patient><id value="patient-two" /><meta><versionId value="bb-king" /></meta><text><status value="additional" /><div xmlns="http://www.w3.org/1999/xhtml">Another great blues player</div></text><active value="true" /><name><use value="nickname" /><family value="King" /></name></Patient></resource><search><mode value="match" /></search></entry></Bundle>
