+++ src/Hl7.Fhir.Core.Tests/TestData/summary/summary-text.json
@@ -1,0 +1,1 @@
+﻿{"resourceType":"Patient","id":"patient-one","meta":{"id":"eric-clapton","versionId":"1234","tag":[{"system":"http://hl7.org/fhir/v3/ObservationValue","code":"SUBSETTED"}]},"text":{"status":"generated","div":"<div xmlns='http://www.w3.org/1999/xhtml'>A great blues player</div>"}}
