+++ src/Hl7.Fhir.Core.Tests/TestData/summary/summary-true.json
@@ -1,0 +1,1 @@
+﻿{"resourceType":"Patient","id":"patient-one","meta":{"id":"eric-clapton","versionId":"1234","tag":[{"system":"http://hl7.org/fhir/v3/ObservationValue","code":"SUBSETTED"}]},"active":true,"name":[{"use":"official","family":"Clapton"}],"gender":"male","birthDate":"2015-07-09"}
