+++ src/Hl7.Fhir.ElementModel/MaskingNode.cs
@@ -20,7 +20,7 @@ namespace Hl7.Fhir.ElementModel
         /// <summary>
         /// Set to true when a complex type property is mandatory so all its children need to be included
         /// </summary>
-        private bool IncludeAll { get; set;}
+        private bool _includeAll { get; set;}
 
         public static MaskingNode ForSummary(ITypedElement node) =>
             new MaskingNode(node, new MaskingNodeSettings
@@ -75,7 +75,7 @@ namespace Hl7.Fhir.ElementModel
         {
             Source = source;
             _settings = parent._settings;
-            IncludeAll = includeAll;
+            _includeAll = includeAll;
             ExceptionHandler = parent.ExceptionHandler;
         }
 
@@ -98,16 +98,12 @@ namespace Hl7.Fhir.ElementModel
 
         public IElementDefinitionSummary Definition => Source.Definition;
 
-        private bool included(ITypedElement node)
+        private (bool included, bool mandatory) included(ITypedElement node)
         {
-            // Trivially, we will include the root
-            if (!node.Location.Contains(".")) return true;
-
             var scope = getScope(node);
 
-            //bool atTopLevel() =>
-            //    // check whether there is a single path separator -> path looks like 'Resource.xxxxx'
-            //    node.LocalLocation.IndexOf('.') == node.LocalLocation.LastIndexOf('.');
+            // Trivially, we will include the root
+            if (!scope.Location.Contains(".")) return (true,false);
 
             bool atRootBundle() => atBundle() && scope.ParentResource == null;
             bool atBundle() => scope.NearestResourceType == "Bundle";
@@ -116,16 +112,19 @@ namespace Hl7.Fhir.ElementModel
             {
                 case MaskingNodeSettings.PreserveBundleMode.All when atBundle():
                 case MaskingNodeSettings.PreserveBundleMode.Root when atRootBundle():
-                    return true;
-                    // else fall through
+                    return (true,false);
+
+                // fall through...
             }
 
-            var included = _settings.IncludeAll || IncludeAll;
+            var included = _settings.IncludeAll || _includeAll;
 
+            bool mandatory=false;         // included because it's required & includeMandatory is on
             var ed = scope.Definition;
             if (ed != null)
             {
-                included |= _settings.IncludeMandatory && ed.IsRequired;
+                mandatory = _settings.IncludeMandatory && ed.IsRequired;
+                included |= mandatory;
                 included |= _settings.IncludeInSummary && ed.InSummary;
             }
 
@@ -134,7 +133,7 @@ namespace Hl7.Fhir.ElementModel
             included |= _settings.IncludeElements?.Any(matches) ?? false;
 
             if (_settings.ExcludeElements?.Any(matches) == true)
-                return false;
+                return (false,false);
 
             bool matches(string filter)
             {
@@ -142,11 +141,10 @@ namespace Hl7.Fhir.ElementModel
                 return loc == f || loc.StartsWith(f + ".") || loc.StartsWith(f + "[");    // include matches + children
             }
 
+            if (_settings.ExcludeMarkdown && scope.InstanceType == "markdown") return (false,false);
+            if (_settings.ExcludeNarrative & scope.InstanceType == "Narrative") return (false,false);
 
-            if (_settings.ExcludeMarkdown && scope.InstanceType == "markdown") return false;
-            if (_settings.ExcludeNarrative & scope.InstanceType == "Narrative") return false;
-
-            return included;
+            return (included,mandatory);
         }
 
 
@@ -159,6 +157,9 @@ namespace Hl7.Fhir.ElementModel
         }
 
         public IEnumerable<ITypedElement> Children(string name = null) =>
-            Source.Children(name).Where(c => included(c)).Select(c => new MaskingNode(this, c, _settings.IncludeMandatory && !Source.Location.Equals("Bundle") ? true : false));
+            from c in Source.Children(name)
+            let inc = included(c)
+            where inc.included
+            select new MaskingNode(this, c, inc.mandatory);
     }
 }
